Implement RUM event sampling and rate limiting for improved performance and resource management
- Introduced a sampling rate for RUM events to reduce database load in production while maintaining statistical significance.
- Updated the RUM POST route to apply sampling logic, allowing for silent drops of non-sampled events.
- Added rate limiting for RUM events, allowing up to 30 unauthenticated requests per minute to enhance performance and prevent abuse.
- Refactored the Guest Event Join Page to utilize server-side validation for event existence, improving security and user experience.

These changes aim to optimize resource usage and enhance the overall performance of the application.
